Initialize asguard devices to NULL in module init
When loading the asguard module, the user specifies which NIC should be
used (by providing the corresponding ifindex of the NIC).

ASGuard then initializes only memory for the specified NIC, other NICs
should be uninitalized if not used.

On ASGuard exit, ASguard iterates through the asguard device array - and
only skips devices if they are a NULL ptr.

This commit fixes that uninitalized devices were not NULL.

Reference Issue #26